requirements.txt
transformers
torch
mysql-connector-python
pandas
flask